requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
chart-studio==1.1.0
click==7.1.1
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.3
pandas==1.0.3
plotly==4.6.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
